chore(environment): tidy up the page and store logic
