+++ resources/js/components/domain/Environment/index.ts
@@ -1,1 +1,3 @@
-export { default as EnvironmentCollectionsManager } from './EnvironmentCollectionsManager.vue';
+export { default as EnvironmentCollectionHeader } from './EnvironmentCollectionHeader.vue';
+export { default as EnvironmentCollectionList } from './EnvironmentCollectionList.vue';
+export { default as EnvironmentVariablesEditor } from './EnvironmentVariablesEditor.vue';
